rss-api/pyproject.toml
[tool.poetry]
name = "rss_api"
version = "0.1.0"
description = ""
authors = ["Romain Dorgueil"]
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.12"
harp-proxy = { path = "../../harp/", develop = true }
feedparser = "^6.0.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
